chore: move withSuspense to new hoc directory
