+++ apps/recnet/src/components/hoc/withSuspense.tsx
